refactor: extract P&L fixture helpers into utils
