+++ src/fixtures/profitAndLoss/mocks.ts
@@ -1,11 +1,9 @@
-import type { LineItem } from '@schemas/common/lineItem'
 import type {
   ProfitAndLoss,
   ProfitAndLossSummary,
 } from '@schemas/reports/profitAndLoss'
 
 import {
-  type CategorySplit,
   COGS_SPLIT,
   INCOME_SPLIT,
   OPEX_SPLIT,
@@ -13,32 +11,16 @@ import {
   PROFIT_AND_LOSS_FIXTURE_START_YEAR,
 } from '@fixtures/profitAndLoss/constants'
 import { schema } from '@fixtures/profitAndLoss/schema'
+import {
+  EMPTY_SUMMARY,
+  makeLineItem,
+  makeSectionLineItem,
+  sumSummaries,
+} from '@fixtures/profitAndLoss/utils'
 import { createGenerator } from '@fixtures/utils/createGenerator'
+import { fromMonthIndex, toMonthIndex } from '@fixtures/utils/monthIndex'
 
 const generate = createGenerator(schema)
-
-const EMPTY_SUMMARY: Omit<ProfitAndLossSummary, 'year' | 'month'> = {
-  income: 0,
-  costOfGoodsSold: 0,
-  grossProfit: 0,
-  operatingExpenses: 0,
-  profitBeforeTaxes: 0,
-  taxes: 0,
-  netProfit: 0,
-  fullyCategorized: true,
-  totalExpenses: 0,
-  uncategorizedInflows: 0,
-  uncategorizedOutflows: 0,
-  uncategorizedTransactions: 0,
-  categorizedTransactions: 0,
-}
-
-const toMonthIndex = (year: number, month: number) => year * 12 + (month - 1)
-
-const fromMonthIndex = (index: number) => ({
-  year: Math.floor(index / 12),
-  month: (index % 12) + 1,
-})
 
 export const makeProfitAndLossSummary = (year: number, month: number): ProfitAndLossSummary => {
   const now = new Date()
@@ -82,55 +64,6 @@ export const makeProfitAndLossSummaries = (
   makeSummariesForMonthIndices(
     toMonthIndex(startYear, startMonth),
     toMonthIndex(endYear, endMonth),
-  )
-
-const makeLineItem = (
-  name: string,
-  displayName: string,
-  value: number,
-  lineItems: LineItem[] = [],
-): LineItem => ({
-  name,
-  displayName,
-  value,
-  isContra: false,
-  lineItems,
-})
-
-const splitIntoLineItems = (total: number, split: readonly CategorySplit[]): LineItem[] => {
-  const values = split.map(([, , share]) => Math.round(total * share))
-  const allocated = values.reduce((sum, value) => sum + value, 0)
-  if (values.length > 0) values[0] += total - allocated
-
-  return split.map(([name, displayName], index) => makeLineItem(name, displayName, values[index]))
-}
-
-const makeSectionLineItem = (
-  name: string,
-  displayName: string,
-  total: number,
-  split: readonly CategorySplit[],
-): LineItem => makeLineItem(name, displayName, total, splitIntoLineItems(total, split))
-
-// Every summary aggregate is linear, so range totals are just field-wise sums.
-const sumSummaries = (summaries: ProfitAndLossSummary[]) =>
-  summaries.reduce(
-    (acc, summary) => ({
-      income: acc.income + summary.income,
-      costOfGoodsSold: acc.costOfGoodsSold + summary.costOfGoodsSold,
-      grossProfit: acc.grossProfit + summary.grossProfit,
-      operatingExpenses: acc.operatingExpenses + summary.operatingExpenses,
-      profitBeforeTaxes: acc.profitBeforeTaxes + summary.profitBeforeTaxes,
-      taxes: acc.taxes + summary.taxes,
-      netProfit: acc.netProfit + summary.netProfit,
-      totalExpenses: acc.totalExpenses + summary.totalExpenses,
-      uncategorizedInflows: acc.uncategorizedInflows + summary.uncategorizedInflows,
-      uncategorizedOutflows: acc.uncategorizedOutflows + summary.uncategorizedOutflows,
-      uncategorizedTransactions: acc.uncategorizedTransactions + summary.uncategorizedTransactions,
-      categorizedTransactions: acc.categorizedTransactions + summary.categorizedTransactions,
-      fullyCategorized: acc.fullyCategorized && summary.fullyCategorized,
-    }),
-    EMPTY_SUMMARY,
   )
 
 export const makeProfitAndLossReport = (
